Fix the left compiler warnings.
1. Resolve all of the ignorance in using signed/unsigned integer types
2. Replace uint64 with time_t when defining variables to store time.
3. Fix one or two type mismatch of bool with integer

+++ src/server/game/AI/SimpleAI.cpp
@@ -222,7 +222,7 @@ void SimpleAI::UpdateAI(uint32 diff)
         if (Spell_Timer[i] < diff)
         {
             //Check if this is a percentage based
-            if (Spell[i].First_Cast < 0 && Spell[i].First_Cast > -100 && me->GetHealth()*100 / me->GetMaxHealth() > -Spell[i].First_Cast)
+            if (Spell[i].First_Cast < 0 && Spell[i].First_Cast > -100 && int32(me->GetHealth()*100 / me->GetMaxHealth()) > -Spell[i].First_Cast)
                 continue;
 
             //Check Current spell
